Make Container types public
Normally, the public API to roaring is the Bitmap; all the container
logic is private. But we want to work with containers, and we don't want
to re-write a bunch of rust code (or copy this code into our crate), so
instead, we fork roaring-rs and make public those types we need.